Allow non-file-based stylesheets.
Instead of just allowing URL-based stylesheet to be copied into the final
output, allow for stylesheet to be provided as local text (CDATA?) inside
the staticsite.xml file.

Alternatively this can be used, with the help of XInclude, to collapse a
series of CSS files into one, final CSS file.