Make youtube-exposing pages export OpenGraph metadata as YouTube.
Unless we give a custom thumbnail on the page, use the one provided by
youtube; also add an og:video property (even though it doesn't seem
described on the OpenGraph page), and expose that in the <head> tag.

Note that the og:video:type property exposed by YouTube cannot be described
as RDF, and it sounds fishy.